renaming python/cxx/sunbeam*.* -> common *.*
