+++ python/cxx/eclipse_config.cpp
@@ -5,7 +5,7 @@
 #include <opm/parser/eclipse/EclipseState/SummaryConfig/SummaryConfig.hpp>
 #include <opm/parser/eclipse/EclipseState/SimulationConfig/SimulationConfig.hpp>
 
-#include "sunbeam.hpp"
+#include "common.hpp"
 
 
 void sunbeam::export_EclipseConfig(py::module& module)
